add delete method

+++ src/main/java/com/example/todo/repositoty/TodoRepository.java
@@ -1,7 +1,8 @@
 package com.example.todo.repositoty;
 import com.example.todo.model.TodoEntity;
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TodoRepository extends JpaRepository<TodoEntity,Long> { }
+public interface TodoRepository extends JpaRepository<TodoEntity,Integer> { }
